Switched or confirmed ground and border for Block1, 2, 3, and 5. Linked room change signals between available block rooms and Village South.
If you encounter problems with changing scenes, look at LevelEntrance.gd (same for puzzle and area scene changes). A signal (enter_or_exit) from a collision area using LevelEntrance.gd can be connected to another LevelEntrance.gd as in the case of the exit area in the ground and border object--the area needs to be tethered to the main node of the scene and it's name needs to match the scene name (e.g., Block1.tscn's owner spatial node must be named "Block1") and the signal must be connected to the enter_or_exit_puzzle method in that spatial node or the transition will not work. Ask MrDudeIII if you have questions.